@@ -1,10 +1,11 @@
-package es.MueblesCastilla.MueblesCastilla.domain.repository;
+package es.MueblesCastilla.MueblesCastilla.domain.userCase;
 
 import java.util.List;
 import java.util.Optional;
+
 import es.MueblesCastilla.MueblesCastilla.domain.dto.UsuarioPojo;
 
-public interface IUsuarioRepository {
+public interface UsuarioService {
 	/**
 	  * Devuelve una lista con todos los usuarios
 	  * @return Lista todos los usuario.
@@ -16,7 +17,7 @@ public interface IUsuarioRepository {
 	 * @return Optional de usuario
 	 */
 	Optional<UsuarioPojo> getUsuarioById(Integer id);//Optional evita la exception nullPointerException
-
+	
 	/**
 	 * Devuelve un usuario dado su email
 	 * @param email de usuario
@@ -24,7 +25,6 @@ public interface IUsuarioRepository {
 	 */
 	Optional<UsuarioPojo> getUsuarioByEmail(String email);//Optional evita la exception nullPointerException
 
-	
 	/**
 	 * Guarda un usuario
 	 * @param newUsuario usuario a guardar
@@ -33,9 +33,18 @@ public interface IUsuarioRepository {
 	UsuarioPojo save(UsuarioPojo newUsuario);
 	
 	/**
+	 * Actualiza un usuario
+	 * @param Recibe un usuario
+	 * @return Devuelve el usuario actualizado
+	 */
+	Optional <UsuarioPojo> update (UsuarioPojo usuario);
+	
+	/**
 	 * Elimina un usuario dado su id
 	 * @param idUsuario Id para eliminar la usuario
 	 */
-	void delete (Integer idUsuario);
+	boolean delete (Integer idUsuario);
 
 }
+
+
