@@ -6,6 +6,6 @@ import es.MueblesCastilla.MueblesCastilla.persistance.entity.DetalleCompra;
 /**
  * DAO DATA ACCESS OBJECT (Objeto accesible a la base de datos)
  */
-public interface IDetalleCompraCrudRepository extends JpaRepository<DetalleCompra, Integer>{
+public interface DetalleCompraCrudRepository extends JpaRepository<DetalleCompra, Integer>{
 
 }
