@@ -20,8 +20,6 @@ import lombok.Setter;
 
 	@Entity
 	@Table(name = "productos")//Para cambiar el nombre en BD
-	@Getter
-	@Setter
 	public class Producto {
 		
 		@Id
@@ -57,6 +55,78 @@ import lombok.Setter;
 			this.descripcion = descripcion;
 			this.idUsuario = idUsuario;
 			this.detalleCompra = detalleCompra;
+			this.usuario = usuario;
+		}
+
+		public Integer getId() {
+			return id;
+		}
+
+		public void setId(Integer id) {
+			this.id = id;
+		}
+
+		public String getNombre() {
+			return nombre;
+		}
+
+		public void setNombre(String nombre) {
+			this.nombre = nombre;
+		}
+
+		public String getImagen() {
+			return imagen;
+		}
+
+		public void setImagen(String imagen) {
+			this.imagen = imagen;
+		}
+
+		public double getPrecio() {
+			return precio;
+		}
+
+		public void setPrecio(double precio) {
+			this.precio = precio;
+		}
+
+		public int getCantidad() {
+			return cantidad;
+		}
+
+		public void setCantidad(int cantidad) {
+			this.cantidad = cantidad;
+		}
+
+		public String getDescripcion() {
+			return descripcion;
+		}
+
+		public void setDescripcion(String descripcion) {
+			this.descripcion = descripcion;
+		}
+
+		public Integer getIdUsuario() {
+			return idUsuario;
+		}
+
+		public void setIdUsuario(Integer idUsuario) {
+			this.idUsuario = idUsuario;
+		}
+
+		public List<DetalleCompra> getDetalleCompra() {
+			return detalleCompra;
+		}
+
+		public void setDetalleCompra(List<DetalleCompra> detalleCompra) {
+			this.detalleCompra = detalleCompra;
+		}
+
+		public Usuario getUsuario() {
+			return usuario;
+		}
+
+		public void setUsuario(Usuario usuario) {
 			this.usuario = usuario;
 		}
 
