@@ -20,8 +20,6 @@ import lombok.Setter;
 
 	@Entity//entidad
 	@Table(name = "usuarios")
-	@Getter
-	@Setter
 	public class Usuario {
 		
 		@Id
@@ -58,7 +56,66 @@ import lombok.Setter;
 			this.productos = productos;
 			this.compras = compras;
 		}
-
+		public Integer getId() {
+			return id;
+		}
+		public void setId(Integer id) {
+			this.id = id;
+		}
+		public String getNombre() {
+			return nombre;
+		}
+		public void setNombre(String nombre) {
+			this.nombre = nombre;
+		}
+		public String getUsername() {
+			return username;
+		}
+		public void setUsername(String username) {
+			this.username = username;
+		}
+		public String getEmail() {
+			return email;
+		}
+		public void setEmail(String email) {
+			this.email = email;
+		}
+		public String getDireccion() {
+			return direccion;
+		}
+		public void setDireccion(String direccion) {
+			this.direccion = direccion;
+		}
+		public String getTelefono() {
+			return telefono;
+		}
+		public void setTelefono(String telefono) {
+			this.telefono = telefono;
+		}
+		public String getTipo() {
+			return tipo;
+		}
+		public void setTipo(String tipo) {
+			this.tipo = tipo;
+		}
+		public String getPassword() {
+			return password;
+		}
+		public void setPassword(String password) {
+			this.password = password;
+		}
+		public List<Producto> getProductos() {
+			return productos;
+		}
+		public void setProductos(List<Producto> productos) {
+			this.productos = productos;
+		}
+		public List<Compra> getCompras() {
+			return compras;
+		}
+		public void setCompras(List<Compra> compras) {
+			this.compras = compras;
+		}
 
 	}
 
