@@ -6,33 +6,28 @@ import java.util.Optional;
 import org.springframework.stereotype.Repository;
 
 import es.MueblesCastilla.MueblesCastilla.domain.dto.UsuarioPojo;
-import es.MueblesCastilla.MueblesCastilla.domain.repository.IUsuarioRepository;
-import es.MueblesCastilla.MueblesCastilla.persistance.crud.IUsuarioCrudRepository;
-import es.MueblesCastilla.MueblesCastilla.persistance.mapper.IUsuarioMapper;
-import lombok.RequiredArgsConstructor;
+import es.MueblesCastilla.MueblesCastilla.domain.repository.UsuarioRepository;
+import es.MueblesCastilla.MueblesCastilla.persistance.crud.UsuarioCrudRepository;
+import es.MueblesCastilla.MueblesCastilla.persistance.mapper.UsuarioMapper;
 /**
  * Repository de Usuario (es el que se conecta a la base de datos DAO)
  */
-@RequiredArgsConstructor//Anotara el constructor con los atributos final.
 @Repository
-public class UsuarioRepository implements IUsuarioRepository{
+public class UsuarioRepositoryImp implements UsuarioRepository{
 	/**
 	 * Crud de Usuario
 	 */
-	private final IUsuarioCrudRepository iUsuarioCrudRepository;
+	private final UsuarioCrudRepository usuarioCrudRepository;
 	/**
 	 * Mapper de usuario
 	 */
-	private final IUsuarioMapper iUsuarioMapper;
+	private final UsuarioMapper usuarioMapper;
 	
-	/*public UsuarioRepository(IUsuarioCrudRepository iUsuarioCrudRepository, IUsuarioMapper iUsuarioMapper) {
+	public UsuarioRepositoryImp(UsuarioCrudRepository usuarioCrudRepository, UsuarioMapper usuarioMapper) {
 		
-		this.iUsuarioCrudRepository=iUsuarioCrudRepository;
-		this.iUsuarioMapper=iUsuarioMapper;
-		
-		****lo creo con lombok****
-	}*/
-	
+		this.usuarioCrudRepository=usuarioCrudRepository;
+		this.usuarioMapper=usuarioMapper;
+	}
 	
     /**
      * Retorna una lista de usuario transformada a pojo de una lista de entidades
@@ -42,7 +37,7 @@ public class UsuarioRepository implements IUsuarioRepository{
 	@Override
 	public List<UsuarioPojo> getAll() {
 		
-		return iUsuarioMapper.toUsuarioPojo(iUsuarioCrudRepository.findAll());
+		return usuarioMapper.toUsuarioPojo(usuarioCrudRepository.findAll());
 	}
 	
 	
@@ -54,7 +49,7 @@ public class UsuarioRepository implements IUsuarioRepository{
 	@Override
 	public Optional<UsuarioPojo> getUsuarioByEmail(String email) {
 		
-		return iUsuarioCrudRepository.findByEmail(email).map(iUsuarioMapper::toUsuarioPojo);//mapea la interfaz mapper y llama al metodo toUsuarioPojo.
+		return usuarioCrudRepository.findByEmail(email).map(usuarioMapper::toUsuarioPojo);//mapea la interfaz mapper y llama al metodo toUsuarioPojo.
 		
 //		return iUsuarioCrudRepository.findById(id).map(hola -> iUsuarioMapper.toUsuarioPojo(hola)); Función lambda(programacion funcional).
 	}
@@ -67,7 +62,7 @@ public class UsuarioRepository implements IUsuarioRepository{
 	@Override
 	public Optional<UsuarioPojo> getUsuarioById(Integer id) {
 		
-		return iUsuarioCrudRepository.findById(id).map(iUsuarioMapper::toUsuarioPojo);//mapea la interfaz mapper y llama al metodo toUsuarioPojo.
+		return usuarioCrudRepository.findById(id).map(usuarioMapper::toUsuarioPojo);//mapea la interfaz mapper y llama al metodo toUsuarioPojo.
 		
 //		return iUsuarioCrudRepository.findById(id).map(hola -> iUsuarioMapper.toUsuarioPojo(hola)); Función lambda(programacion funcional).
 	}
@@ -79,7 +74,7 @@ public class UsuarioRepository implements IUsuarioRepository{
 	@Override
 	public UsuarioPojo save(UsuarioPojo newUsuario) {
 		
-		return iUsuarioMapper.toUsuarioPojo(iUsuarioCrudRepository.save(iUsuarioMapper.toUsuarioEntity(newUsuario)));
+		return usuarioMapper.toUsuarioPojo(usuarioCrudRepository.save(usuarioMapper.toUsuarioEntity(newUsuario)));
 	}
 	/**
 	 * Elimina usuario
@@ -88,7 +83,7 @@ public class UsuarioRepository implements IUsuarioRepository{
 	 */
 	@Override
 	public void delete(Integer idUsuario) {
-		iUsuarioCrudRepository.deleteById(idUsuario);
+		usuarioCrudRepository.deleteById(idUsuario);
 		
 	}
 
